tests/test_mvp.py
from __future__ import annotations

import sys
from pathlib import Path

import yaml

# Ensure repo root is importable even without `pip install -e .`
ROOT = Path(__file__).resolve().parents[1]
sys.path.insert(0, str(ROOT))

from scopebench.contracts import TaskContract  # noqa: E402
from scopebench.plan import PlanDAG  # noqa: E402
from scopebench.runtime.guard import evaluate  # noqa: E402


def load_yaml(path: Path):
    return yaml.safe_load(path.read_text(encoding="utf-8"))


def test_phone_overreach_denied():
    contract = TaskContract.model_validate(load_yaml(ROOT / "examples/phone_charge.contract.yaml"))
    plan = PlanDAG.model_validate(load_yaml(ROOT / "examples/phone_charge.plan.yaml"))
    res = evaluate(contract, plan)
    assert res.policy.decision.value == "DENY"


def test_swe_fix_allowed():
    contract = TaskContract.model_validate(load_yaml(ROOT / "examples/swe_fix.contract.yaml"))
    plan = PlanDAG.model_validate(load_yaml(ROOT / "examples/swe_fix.plan.yaml"))
    res = evaluate(contract, plan)
    assert res.policy.decision.value == "ALLOW"


def test_ops_rotate_key_asks():
    contract = TaskContract.model_validate(load_yaml(ROOT / "examples/ops_rotate_key.contract.yaml"))
    plan = PlanDAG.model_validate(load_yaml(ROOT / "examples/ops_rotate_key.plan.yaml"))
    res = evaluate(contract, plan)
    assert res.policy.decision.value in {"ASK", "DENY"}  # enterprise preset should at least ASK for IAM/prod


def test_scope_laundering_detected_by_aggregation():
    # Each step says "small", but repeated dependency creation should accumulate.
    contract = TaskContract.model_validate(
        {
            "goal": "Do a small tidy-up",
            "preset": "personal",
            "thresholds": {
                "max_dependency_creation": 0.55,
                "max_power_concentration": 0.25,
            },
        }
    )
    plan = PlanDAG.model_validate(
        {
            "task": "Do a small tidy-up",
            "steps": [
                {"id": "1", "description": "Create service A (small).", "tool": "analysis"},
                {"id": "2", "description": "Create service B (small).", "tool": "analysis", "depends_on": ["1"]},
                {"id": "3", "description": "Create service C (small).", "tool": "analysis", "depends_on": ["2"]},
            ],
        }
    )
    res = evaluate(contract, plan)
    # We don't assert DENY because MVP heuristics are simple; but we expect not silently ALLOW.
    assert res.policy.decision.value in {"ASK", "DENY"}
