pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scopebench"
version = "0.1.0"
description = "Plan-level proportionality enforcement (scope alignment) for tool-using agents."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "ScopeBench contributors"}]
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[project.scripts]
scopebench = "scopebench.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["scopebench*"]

[tool.ruff]
line-length = 100
target-version = "py310"
